Overhaul Perl5::TestEachCommit::Util and its tests
Attempting to use Perl5::TestEach::Commit::Util::process_command_line()
in an outside program demonstrated that it was not calling GetOptions()
correctly, even though it was passing its tests in t/02-util.t.

In the course of working on this, I decided to discontinue
Perl5::TestEach::Commit::Util::usage() for the time being.