Within examine_one_commit(), set $ENV{PERL_CORE} true.
Suppose a test of a CPAN module behaves differently when it knows it's
being run inside the Perl 5 core distribution rather than anywhere else.
Example: cpan/Term-Table/t/honor_env_in_non_tty.t (line 6):

  6    if (!$ENV{PERL_CORE}) {
  7         print STDERR "# Using Test2::Tools::Tiny " . Test2::Tools::Tiny->VERSION . "\n";
  8    }
  9    else {
 10         print "# Using Test2::Tools::Tiny\n";
 11    }

In the Perl core, the message is printed to STDOUT -- which in *this*
library we are often redirecting to /dev/null, as we're only interested
in the PASS/FAIL status of each stage.  So we need to tell
examine_one_commit(), "Behave as if you're being invoked within the Perl
core, even when you're not."