workdir: Should check for truth, not simply definedness
During testing in an outside program, new() was provided with KVP
"workdir => ''", i.e., Perl-false but defined.  Failing to check for
truth meant that the program did not proceed to check for
SECONDARY_CHECKOUT_DIR and was content to assign '' to workdir inside.
But that immediately failed the -d $args->{workdir} check.